Balance tweaks
Start params: decrease spawnMaxGap so asteroids are more frequent at the
start. Before, I found that I could go without moving at all for quite a
bit and not hit anything. This change gives the player a bit more to do
without making it too challenging right off the bat.

Decrease spawnPUEvry and delayPerUnitOfDistance for end params to make
sure delay doesn't get too large - with the settings in this commit, I
find that the delay maxes out at a point where you have to issue your
command as soon as the asteroid comes on screen to avoid hitting it, so
I think this is as high as we want these to be.

Increase spawnMinGap for end params - I think gaps of less than .3 feel
too frustrating given the imprecise feel of the controls.

To help compensate for difficulty decreases for end state and to create
a more chaotic feel, decreased spawnMaxGap.